backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
google-genai==0.8.0
pymupdf==1.24.14
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
cryptography==43.0.3
pydantic-settings==2.6.1
pydantic==2.10.3
python-multipart==0.0.12
email-validator==2.2.0
httpx==0.28.1
anyio==4.7.0
